got phase 1 of galactic generator working
Needs heavy optimization. As is, it currently generates an average of
~24,672,376 lens flares when you can see the whole galaxy, which is
obviously problematic.

However, it is generating stars based on the camera's view, which is
cool.

Next up:
-Choose from a set of prefabs that includes stars, star clusters,
nebulae, etc. when generating. (must remember to enable mipmaps for
those guys)
-After that is combining some of those objects as they get farther away.

THAT will be the hard part. Probably going to involve Rendertextures,
which I don't know how to use yet.